% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalise_filter_counts}
\alias{normalise_filter_counts}
\title{Normalisation}
\usage{
normalise_filter_counts(counts)
}
\arguments{
\item{counts}{The counts matrix, with features in columns}
}
\description{
Normalisation
}
